Resolve #965 bug - polyline having too many points duplicated, and #933, moving converts to query
